Refactor range_lock in fs
